% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness.check.R
\name{robustness.check}
\alias{robustness.check}
\title{Bayes Factor Robustness Check}
\usage{
robustness.check(
  x,
  y,
  mu = 0,
  nullInterval = NULL,
  paired = FALSE,
  max_r = 2,
  iterations = 200
)
}
\arguments{
\item{x}{a vector of observations for the first (or only) group}

\item{y}{a vector of observations for the second group (or condition, for paired)}

\item{mu}{used to specify test value for one-sample designs}

\item{nullInterval}{optional vector of length 2 containing lower and upper bounds of an interval hypothesis to test}

\item{paired}{if TRUE, observations are paired}

\item{max_r}{maximum value of the scale parameter of the Cauchy prior to be plotted}

\item{iterations}{will determine how precise the plot will be}
}
\description{
This function plots the change in the Bayes Factor under different possible
priors. For an example, see the BayesianTTestTutorial vignette.
}
\author{
Roy Michael Moore, \email{roy.moore@student.uva.nl}
}
