% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness_check.R
\name{robustness_check}
\alias{robustness_check}
\title{Bayes Factor Robustness Check}
\usage{
robustness_check(x, y, paired = FALSE, max_r = 2, iterations = 200)
}
\arguments{
\item{x}{a vector of observations for the first (or only) group}

\item{y}{a vector of observations for the second group (or condition, for paired)}

\item{paired}{if TRUE, observations are paired}

\item{max_r}{maximum value of the scale parameter of the Cauchy prior to be plotted}

\item{iterations}{will determine how precise the plot will be}
}
\description{
This function plots the change in the Bayes Factor under different possible priors
}
\author{
Roy Michael Moore, \email{roy.moore@student.uva.nl}
}
