% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.calculator.R
\name{ttest.calculator}
\alias{ttest.calculator}
\title{Bayesian t-test calculator App}
\usage{
ttest.calculator()
}
\description{
This function plots the change in the Bayes Factor under different possible priors
}
\author{
Roy Michael Moore, \email{roy.moore@student.uva.nl}
}
